datastore-api-server/requirements3.txt
click==6.6
connexion==1.0.109
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
psycopg2==2.6.2
PyYAML==3.11
requests==2.11.0
six==1.10.0
SQLAlchemy==1.0.14
strict-rfc3339==0.7
swagger-spec-validator==2.0.2
Werkzeug==0.11.10
